+++ CHANGELOG.md
@@ -20,8 +20,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 
 - `--debug` didn't work because of a tiny error with `modules/log.py` (1.0.0 -> 1.0.1)
-- `config.py 1.0.1`: Various cleanup and fixes
- 
+- `config.py 1.0.1`: Various cleanup and fixes.
+- `exclusion.py 1.0.1`: Various cleanup. More error handling.
+- `inclusion.py 1.0.1`: Various cleanup. More error handling.
+
 ## [2.0.0] - 2025-07-16
  
 ### Added